+++ R/unnest_tokens.R
@@ -16,7 +16,9 @@
 #' @param format Either "text", "man", "latex", "html", or "xml". If not text,
 #' this uses the hunspell tokenizer, and can tokenize only by "word"
 #'
-#' @param to_lower Whether to turn column lowercase.
+#' @param to_lower Whether to convert tokens to lowercase. If tokens include
+#' URLS (such as with \code{token = "tweets"}), such converted URLs may no
+#' longer be correct.
 #'
 #' @param drop Whether original input column should get dropped. Ignored
 #' if the original input and new output column have the same name.
@@ -214,6 +216,10 @@ unnest_tokens.data.frame <- function(tbl, output, input, token = "words",
   ret[[output]] <- unlist(output_lst)
 
   if (to_lower) {
+    if (!is.function(token))
+      if(token == "tweets") {
+        message("Using `to_lower = TRUE` with `token = 'tweets'` may not preserve URLs.")
+      }
     ret[[output]] <- stringr::str_to_lower(ret[[output]])
   }