+++ NEWS.md
@@ -3,6 +3,7 @@
 * Fix bug in `augment()` function for stm topic model.
 * Warn when tf-idf is negative, thanks to @EmilHvitfeldt (#112).
 * Switch from importing broom to importing generics, for lighter dependencies (#133).
+* Add functions for reordering factors (such as for ggplot2 bar plots) thanks to @tmastny (#110).
 
 # tidytext 0.2.0